+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - volume support for API300
   - volume_template support for API300
   - enclosure support for API300. Also added `:patch` action
+  - Added support to SAS Logical Interconnects for API300::Synergy
   - logical_switch support for API300
   - interconnect support for API300
   - server_hardware support for API300. Also added `:patch` action